spark-tensorflow/trainer/requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
funcsigs==1.0.2
html5lib==0.9999999
Markdown==2.2.0
mock==2.0.0
numpy==1.13.0
pbr==3.0.1
protobuf==3.3.0
six==1.10.0
tensorflow==1.2.0
Werkzeug==0.12.2
google-cloud-core==0.27.1
google-cloud-storage==1.4.0
